Single out README details that only apply to XGBoost.
Originally intended to refactor this to have it all in the XGBoost package README, but a lot of the details deal with Colab runtimes so they likely apply to all the notebooks. We likely need to refactor more of the notebooks themselves in order to extract shared logic into libraries, then the documentation can be streamlined further.