Shorten and disambiguate directory names.
The "boost" in "xgb-boost" is redundant, since that's what the "b" stands for. The technique of Kriging interpolation is also just known as "kriging", so the shorter name is already fully specific. "Linear" is a very overloaded term whose meaning can vary a lot even within specific contexts, "regression" has similar problems, but "LRM" is a standard abbreviation for "linear regression model".